pyproject.toml
[project]
name = "tutor"
version = "0.1.0"
description = "MalayChat - Free Local Malay Language Learning App"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.40.0",
    "transformers>=4.46.0",
    "torch>=2.5.0",
    "accelerate>=1.1.0",
    "sentencepiece>=0.2.0",
    "llama-index-core>=0.12.0",
    "llama-index-llms-llama-cpp>=0.4.0",
    "huggingface-hub>=0.27.0",
]

[project.scripts]
malaychat = "main:main"
